Refactor utils tests #7
